requirements.txt
flask~=3.1.0
google-generativeai
requests~=2.32.3
beautifulsoup4~=4.12.3
protobuf~=5.28.3
urllib3
pypdf
fuzzywuzzy